Separate into subcommands and open up for addon library authors
